Hide the yxtFooter when CFilters is active (#630)
Hide the YxtFooter element when CFilters is active

Previously, we were just hiding the answers footer (which contains the yext logo and location info). We also need to hide the YxtFooter so that it doesn't show up on the CFilters page.

Solve this problem by hiding the YxtFooter similarly to how the answers footer is hidden, which is done by using interactions.js to toggle CSS classes on the specific elements.

The breakpoints are needed so that the element is not hidden when someone enters the filters page and then exits the filters page by dragging the window larger. The maps page is different in that the footer is hidden at a smaller breakpoint than the vertical standard or grid layouts.

J=SLAP-1006
TEST=manual

Override the layouts/footer.hbs and add some content. Confirm that the content is hidden when the CFilters page is open. Also test that dragging the brower window to make it larger and therefore disabling the CFilters page makes the YxtFooter show up again. Test both vertical-grid and vertical-map.