interactive-map: Fix cluster click moves to wrong center
Previously, if you clicked on a pin cluster on a mapbxo map, it would
take you to the current center of map, except zoomed in. This behavior
was a result of conflicting setZoom/setCenter calls on the map.

Instead, we want Mapbox to set the zoom center in one call. It turns
out Consulting fixed this in a recent commit, add these changes to our
Consulting map code to allow Mapbox to setZoomCenter in its own way that
does not give wonky behavior.

J=SLAP-1134
TEST=manual

Test that clicking on a mapbox pin cluster will zoom you and take you to
the center of the pins within that cluster. The behavior before was
obviously taking you to the wrong location on a pin cluster click.