interactive-map: Add canvas click behavior
When you click on a pin or a result card, it is meant to show the card
and the pin in a selected state. It is desired behavior that when you
click on the map canvas the current pin and result card should not be
focused anymore.

Because Mapbox and Google handle click events separately, we have to
have provider-specific code for finding when a canvas is clicked.

J=SLAP-1086
TEST=manual

Test that you can click on a pin and then click anywhere on the map
canvas to remove the select styling on both the results list and the
pin.

On mobile, show that clicking on the map canvas when a map pin is
focused will close the detail card and move the map to full height.

Test that you can click on a pin and then directly another pin and these
changes do not cause any unintended behvaior.

Test that you can click on a result card and then click the map canvas
and both the pin and card remove their focused styling.

Test that clicking on the map control or any other pin element on the
map will not trigger a canvas click event.