interactive-map: Remove results-specific storage listeners
Previously, we would keep listeners for card clicks in storage all
searches. This is a problem because part of the listener is relying on
the index of the pinned card remaining static. However, we cannot rely
on this because the index of cards can change with every search.

We remove any results-specific listeners when the map is updated with
new pins. That way, we do not keep our old listeners from pins in a
different search state.

J=None
TEST=manual

Test that you can click on a cluster pin and it then click on a card in
the new results list. This should highlight the corresponding pin on the
map.