interactive-map: Add selected cluster pin styling on card click (#639)
This PR is split into two commits: One is adding changes to the
Consulting map code from their latest commits. The second is
building on top of that for our personal use case.

Previously, if you were using pin clustering, if you clicked on a result
card associated with a pin in a pin cluster, the map pin cluster would
not have selected styling. By default, this styling changes the color of
the cluster pin to separate it from other pins.

This change is important because we removed ordinals from cards. It is
important for users to be able to see that the card is associated with a
pin that is not on the map but rather in a cluster.

We build on top of Consulting locator code to store the id of the pins
and find the cluster if it exists.

J=SLAP-1085
TEST=manual

Test that you can click on a card in the results list and have it
highlight a cluster pin to indicate that the card exists in the cluster.

Test that you can click out of the selected cluster pin by clicking on
another card in the results list or on the canvas.

Test on both Google and Mapbox.