interactive-map: Add Interactive Map Page Template and Component (#643)
Add the interactive map page template and component.

The interactive map is meant to be a page that allows for
a user to interact with both the results list and the map and
see behaviors across the whole page. For example, clicking
on a map pin will highlight the corresponding card. Clicking on
a result card will highlight a pin. If the pin is not in the current
viewable area, the map will move to show the pin.

We add two new components: A NewMap (to be re-named)
which is meant to be a wrapper component over Consulting
code to set up a map in the theme. This should be a general
interface. Consulting code allowed us to have advanced features
like pin clustering off the bat.

We also add the InteractiveMap component (to be re-named
to InteractiveMapOrchestrator). This component handles the
interactions between the results, the alternative verticals,
the New Map component, and new Search This Area
and mobile toggles features.

The page html.hbs template also differs a bit to allow
for the new features on the page.

J=PB-10500,PB-10814
TEST=manual

General smoke test that a site can build and the basic interactions
like clicking on a pin, clicking on a cluster, configuration for the start
coordinate, and looking at a mobile detail card works