interactive-map: Fix uncommenting navigation component (#645)
We should be able to uncomment the navigation partials in the page hbs
for vertical-interactive-map. This should show the navigation component
with the correct styling. Fix a bug where we were referencing the wrong
partial. Then fix the styling so it stays within its container.

J=SLAP-1166
TEST=manual

Test that you can see the Navigation component when you uncomment the
script partial and the html partial. Also ensure that the Navigation
does not leak its container horizontally, on the latest theme.