scripts: Use canary/latest css and templates
While we are in EA, we want to use the canary/latest branch of the SDK
to see how our changes work with the new library changes. This means we
use canary/latest for all SDK assets.

This item is necessary because we noticed that the star icon was still
showing for verticals on univeral pages. Specifically, the star icon
would show when this vertical was not an entry in any verticalsToConfig
of any config.json file.

J=SLAP-1173
TEST=manual

Run on a local jambo site, show that stars aren't showing for verticals
that have no corresponding entry in any verticalsToConfig.