vertical-full-page-map: z-index fixes for android (#662)
Explicitly set the z-index for map controls as lower than the z-index
for the results list. There was a bug where the map controls (Mapbox
logo, the zoom controls) would show on the results list. By explicitly
setting the results list as higher (and subsequently the map/list toggle
higher than that), we don't see that behavior anymore on mobile.

J=SLAP-1150
TEST=manual

Test on Browserstack that when you look at a vertical full page map
results list on mobile, you do not see the zoom controls anywhere on the
landing page

Tested on a Samsung S21 on Google Chrome specifically. Also tested on
Chrome on MacOS.