Update dependencies to address vulnerabilities
These are dependencies pointed out by dependabot by the hitchhiker
theme solution templates.

In the theme's package.json
- lodash 4.17.15 -> 4.17.20
this was done by adding lodash as a direct devDependency
The theme's package.json is technically unused by the solution templates.
But we might as well update our dev dependency to make
the dependabot warning go away.

In the theme's static/package.json, which is the one directly
used by the solution templates
- underscore.string 3.2.3 -> 3.3.5, also moved this from devDependencies to
  a regular dependency because it's used in the browser
- lodash 4.17.15 -> 4.17.20, this updates lodash in multiple @babel/ dependencies

J=SLAP-1196
TEST=manual

test that I can load up a map page (regular vertical-map page not the
vertical-full-page-map) on ie11, and the address will be displayed correctly