Allow custom cards to be used with the `vertical` command.
The `vertical` command had a bug that caused it to only allow built-in cards
to be specified as the `cardName` flag. This was overly restrictive. Users of
the command should be able to configure their vertical to use one of their
custom cards as well.

J=SLAP-1188
TEST=manual

Made a test site with custom cards. Verified that the card names appeared in
the `describe` output for `vertical`. Also verified that I could pass one
of the custom card names to the `cardName` flag and that the vertical would
be configured to use it properly.