vertical-full-page-map: allow specification of clientId in config (#664)
clientId is a config option that can be specified for google maps
for authentication, as an alternative to the api key.
If both an api key and client id are passed to the google maps
script, the api key will take higher priority. Even if no apiKey
is explicitly provided in the page config, we automatically provide
a default, so this commit makes it so if a clientId is provided,
we no longer provide a default map api key.

J=SLAP-1114
TEST=manual,auto

test that I can specify a clientId. see that it will be used to load
in google maps (see "gme-yextinc" work and a bogus clientId not work)

see that I can still specify a custom api key
see that an apiKey is still provided by default it no clientId is specified