vertical-full-page-map: Handle out of bounds longitude
Before, if you moved the map outside of the boundaries of longitude
[-180, 180], it was possible to get outside of the range. This would
cause liveapi to error because the longitude was incorrect. Here we mod
our longitudes so it wraps from -180 to 180 and stays within range.

Note: we do not need to do this for latitude because our current map
implementation does not let you wrap along the latitude lines.

J=SLAP-1107
TEST=manual

On a vertical full page map, zoom out to the maximum value and drag the
map to the left. Make sure when you're at the -180 longitude mark, if
you drag more to the left, you will be in the positive range and
correctly a liveapi request without an error.

Test this the positive longitude direction as well.