vertical-full-page-map: Show navigation by default (#673)
Before, the navigation component was commented out by default. Change
this so it is uncommented by default. This is from requests from the
HHs.

Also change the default height of the mobile map to handle the change
in search header height on mobile.

J=SLAP-1189
TEST=manual

On new a site, creating a page (below) should show the navigation
component below the searchbar by default on both mobile and desktop

Make sure the zoom controls on the map in mobile are still visible (as
the height of the search header is now taller).

e.g. npx jambo page --template vertical-full-page-map --name locations2