vertical-full-page-map: location card click listener only on desktop
Previously, if you were on a mobile device and you clicked on a card,
the "selecting" behavior from desktop would occur. That is, the card
would enter its selected state, which is opening a detail card. This is
not correct because the only entry to a selected result should be
through a pin click on the map view when on mobile.

To fix this, we only enact the selected listeners on card click when we
are on desktop.

J=SLAP-1202
TEST=manual

Test that on mobile/tablet
* Clicking on a result card does not hide the card anymore
* Clicking on a pin in the map view will show a detail card

Test that on desktop
* Clicking on a result card still selects the correct pin (the selected
  styling on the pin shows).
* Clicking on a result card also selects a cluster pin if applicable.