Intelligent defaults for `vertical` command's `cardName`. (#681)
This PR makes the `cardName` defaulting logic more intelligent. Initially,
if no `cardName` was supplied, a default of `standard` was used. But, all
vertical templates prescribe their own default `cardType`. This is now used
as the default. If the template has no `cardType` (which is not currently
the case), a default of `standard` will still be used.

A validation layer was added for the `vertical` command, on top of the
structural validation done already by YArgs. This layer will throw UserErrors
if certain conditions are not met. It allows the other code in the class to
make certain assumptions safely.

J=SLAP-1186
TEST=manual

Tested the following:
- If an invalid `cardName` is provided, an error is thrown.
- If an invalid `template` is provided, an error is thrown.
- If a valid `cardName` is specified, that appears in the new vertical page's
  config as the `cardType`.
- If no `cardName` is specified, but the template has a default `cardType`,
  that is used over `standard`.
- If no `cardName` is specified and the template has no default `cardType`,
  then `standard` is used.