Add a `locales` flag to the `vertical` command. (#686)
This PR adds the `locales` flag to the `vertical` command. This flag is piped
through to the `page` command that is ultimately run. If specified, localized
copies of the vertical's page will be created.

As a side note, this PR further highlights the need for Jambo to export it's
ArgumentMetadata model. We are effectively having to maintain an identical
copy of it in the Theme because we don't have access. Moving the `page` command
to the Theme would have also helped here. The method for obtaining the valid
page locales could then be extracted to a common helper.

J=SLAP-1185
TEST=manual

Tested the following:

- The output of `npx jambo describe` is correct.
- The output of `npx jambo vertical --help` is correct.
- Running the command with no locales works as expected.
- Running the command with an invalid additional locale works as expected.
- Running the command with valid, additional locales works as expected.