vertical-full-page-map: Keep height on map view fixed (#685)
Previously, on mobile phone browsers with navigation bars, we would see
that 100vh would not necessarily be the 100% of the viewport visible to
the user. For instance, on Chrome mobile for iOS, 100vh is obscured by
about 60px to make room for the Chrome navigation tools. This is not a
problem in the mock-mobile on Chrome desktop.

We keep the height of the map view fixed when it is in view. That way,
you are always seeing the important parts of the map. and cannot scroll
away from it. Because of this change, we must fix a few other things,
including the detail card and the wrapper of the map (so we cannot
invisibly scroll to the hidden footer).

NOTE: The Answers logo footer and the customer footer will no longer be
visible on the map view. They will only be visible on the list view.

J=SLAP-1207
TEST=manual

Test on physical iOS device (iPhone 8+) Chrome Safari
Browserstack iPhone 12 Chrome, Safari
Test on Chrome, Firefox, Safari locally on MacOS

Browserstack s21 Chrome, Firefox will NOT work currently, because there
is a bug with the Navigation component styling causing an overflow.

You scroll down the results list and see the Answers logo footer and
custom footer.

You can click the map toggle and see the Search This Area and Location
Bias immediately. It is not obscured by any navigation tools or cut off
in the viewport. You cannot scroll in the map view.

You can click on a pin and see its detail card with the correct height.