Keyboard support for "Search When Map Moves" (#678)
Make the "Search When Map Moves" feature work with the keyboard

Previously the "Search When Map Moves" feature only worked with the mouse because it was relying on the `dragEndHandler` rather than the `panHandler`. By switching to the `panHandler`, the feature will work with the keyboard. As a result of this change, I split up the denouncer so that zoom changes and map center changes are debounced separately.

Also, after a user runs a search, or after a user focuses on a pin which is at the edge of the screen, we don't want to run a new search. To prevent this, I created `PanTriggers` which are analogous to `ZoomTriggers` so that we could prevent running new searches in these cases.

The Google Maps 'idle' event is called repeatedly when the map is moves with the keyboard, which I believe is a bug, so in order to get around this I used a debouncer so that Google Maps only fires the `panHandler`once the map is idle for 100ms.

J=SLAP-1141
TEST=manual

Test on both Mapbox and on Google Maps that "Search When Map Moves" works when navigating the map with the keyboard. Confirm that running a search or focusing on a pin near the edge of the screen doesn't run a new search. Confirm that clicking on a cluster does run a new search after zooming in. Test that I can tab through results without new searches being ran