Have `vertical` command set `verticalKey` in the `defaultInitialSearch`. (#688)
Recently, Rose merged a change that uncommented the `pageSettings.search` block
in all vertical page templates. This was so that pages built with these
templates have a `defaultInitialSearch` set by default. The block also contains
a `verticalKey`. This key needs to be configured properly by the `vertical`
command.

To suppor this, the `_configureVerticalPage` command was altered somewhat. Now,
instead of immeadiately invoking `parse` on the page configuration, some regexs
are run on the raw config first. The regex replaces all instances of
`<REPLACE ME>` with the `verticalKey`. By using a regex, then calling `parse`,
we will not need to update the `vertical` command in the future if additional
`verticalKey` placeholders are added to the template page configs.

J=SLAP-1187
TEST=manual