Fix the vertical full page map zooming out quickly bug (#684)
Fix a bug that incorrectly caused the map the zoom back in while
zooming out

The issue was that the setting and the deleting of `LOCATOR_FROM_SEARCH_THIS_AREA` would sometimes get out of sync when zooming out of the map quickly. This caused the map to automatically update the zoom, which made it difficult to zoom out in some locations because the map would keep zooming back in automatically.

This PR solves this by keeping track of the number of concurrent calls of `searchThisArea`, and only updating the map zoom if that value is equal to zero.

J=SLAP-1142
TEST=manual

Test zooming out quickly and confirm that when the the map does not automatically zoom back in in situations where it would previously occur. Test that performing a search correctly updates the zoom automatically.