Only collapse filters on searchbar searches
This fixes a regression where collapsible filters were
collapsing whenever any search was run, not just ones made
using the searchbar. This is done by checking for the query trigger
corresponding to the search.

This happened because of the way collapsibe filters was differentiating
between vertical results updated caused by the searchbar, and vertical
results updated made by filter components and pagination. Previously,
any update to vertical results that had a corresponding update to query
was considered to be caused by the searchbar. In SDK 1.8, this is no longer
the case, and all searches now have a corresponding query update.

J=SLAP-1204
TEST=manual

test that updating filters in collapsible filters will not collapse the filters
test that back navigating on the filters view (when the filters are expanded)
does not collapse the filters