Adjust the locator debouncer for mobile (#693)
Adjust the locator debouncer so that you don't have to drag so far on mobile to trigger a new search

Decrease the relative distance threshold from .2  to .125

Previously, the relative distance didn't properly take into account mobile screen widths. This updated debouncer uses the window.innerWidth so that the relative distance is accurate regardless of the screen width.

J=SLAP-1210
TEST=manual

Test on various screen widths and print out the relative distance to ensure that the calculations are correct. Test that search on map move triggers on mobile, tablet, and desktop screen sizes with reasonable changes in the distance