vertical-full-page-map: Fix styling for iFrame experience
Previously, the map would resize indefinitely when iframed. This is
because the library we use, the iFrame resizer, listens for changes to
the height of the content in the child frame. In turn, the child frame
is listening for changes to the height in its frame. This is because we
are relying on 100vh heights for the map to give a full screen effect.

This is a know bug (below) with the iframe resizing library that is used
with all of our current iframe experiences.

iframe-resizer/issues/ 692

In order to fix this, we set an explicit height for the maps on mobile
and desktop. This allows the content to not rely on its viewport height.
We set the heights to defaults that look good across many mobile
browsers and desktop browsers.

NOTE: As a result of this, we had to change our result list from being
very long on mobile to being inner scrolled. This has a UI effect of two
scroll bars.

J=SLAP-1201
TEST=manul