Fix behavior when clicking away from detail cards (#695)
Fixes the problem that occurs when a users clicks away from a mobile detail card

Previously, the behavior when a user closed a detail card by clicking on the map and by clicking the close card button. This PR updates the behavior so that both methods of closing the detail card results in the same behavior. This fixes the bug where the detail card would be present below the results when the users switches back to list view

J=SLAP-1205
TEST=manual

Verify that the bug is no longer present. Test that closing a card by clicking on the map and by clicking on the 'x' button results in the same behavior.